+++ it/server/src/test/java/com/walmartlabs/concord/it/server/EscapeGitCommitMessageIT.java
@@ -1,0 +1,102 @@
+package com.walmartlabs.concord.it.server;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.walmartlabs.concord.client.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.HashMap;
+import java.util.Map;
+
+import static com.walmartlabs.concord.it.common.ServerClient.assertLog;
+import static com.walmartlabs.concord.it.common.ServerClient.waitForCompletion;
+import static java.util.Collections.singletonMap;
+
+public class EscapeGitCommitMessageIT extends AbstractServerIT {
+
+    private MockGitSshServer gitServer;
+    private int gitPort;
+
+    @Before
+    public void setUp() throws Exception {
+        Path data = Paths.get(EscapeGitCommitMessageIT.class.getResource("escapeCommitMessage").toURI());
+        Path repo = GitUtils.createBareRepository(data, "oops ${booom}");
+
+        gitServer = new MockGitSshServer(0, repo.toAbsolutePath().toString());
+        gitServer.start();
+
+        gitPort = gitServer.getPort();
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        gitServer.stop();
+    }
+
+    @Test
+    public void test() throws Exception {
+        String orgName = "Default";
+
+        // ---
+
+        String projectName = "project@" + randomString();
+        String repoSecretName = "repoSecret@" + randomString();
+        String repoName = "repo@" + randomString();
+        String repoUrl = String.format(ITConstants.GIT_SERVER_URL_PATTERN, gitPort);
+
+        // ---
+
+        SecretOperationResponse response = generateKeyPair(orgName, repoSecretName, false, null);
+
+        // ---
+
+        RepositoryEntry repo = new RepositoryEntry()
+                .setName(repoName)
+                .setUrl(repoUrl)
+                .setBranch("master")
+                .setSecretId(response.getId());
+        ProjectsApi projectsApi = new ProjectsApi(getApiClient());
+        projectsApi.createOrUpdate(orgName, new ProjectEntry()
+                .setName(projectName)
+                .setRepositories(singletonMap(repoName, repo)));
+
+        // ---
+        Map<String, Object> input = new HashMap<>();
+        input.put("org", orgName);
+        input.put("project", projectName);
+        input.put("repo", repoName);
+        StartProcessResponse spr = start(input);
+
+        // ---
+
+        ProcessApi processApi = new ProcessApi(getApiClient());
+        ProcessEntry psr = waitForCompletion(processApi, spr.getInstanceId());
+
+        // ---
+
+        byte[] ab = getLog(psr.getLogFileName());
+        assertLog(".*Hello, Vasia.*", ab);
+    }
+}